% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fit}
\alias{fit}
\alias{transform}
\title{Generic Method to Train an H2O4GPU Estimator}
\usage{
fit(object, ...)

transform(object, ...)
}
\arguments{
\item{object}{The h2o4gpu model object}

\item{...}{Additional arguments (unused for now).}
}
\description{
Generic Method to Train an H2O4GPU Estimator

Generic Method to Transform a Dataset using Trained H2O4GPU Estimator
}
