% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{h2o4gpu.random_forest_classifier}
\alias{h2o4gpu.random_forest_classifier}
\title{Random Forest Classifier}
\usage{
h2o4gpu.random_forest_classifier(n_estimators = 10L, criterion = "gini",
  max_depth = 3L, min_samples_split = 2L, min_samples_leaf = 1L,
  min_weight_fraction_leaf = 0, max_features = "auto",
  max_leaf_nodes = NULL, min_impurity_decrease = 0,
  min_impurity_split = NULL, bootstrap = TRUE, oob_score = FALSE,
  n_jobs = 1L, random_state = NULL, verbose = 0L, warm_start = FALSE,
  class_weight = NULL, subsample = 1, colsample_bytree = 1,
  num_parallel_tree = 1L, tree_method = "gpu_hist", n_gpus = -1L,
  predictor = "gpu_predictor", backend = "auto")
}
\description{
Random Forest Classifier
}
