% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{h2o4gpu}
\alias{h2o4gpu}
\alias{h2o4gpu-package}
\title{h2o4gpu in R}
\description{
h2o4gpu in R
}
\examples{
\dontrun{

library(h2o4gpu)

# Setup dataset
x <- iris[1:4]
y <- as.integer(iris$Species) - 1

# Initialize and train the classifier
model <- h2o4gpu.random_forest_classifier() \%>\% fit(x, y)

# Make predictions
predictions <- model \%>\% predict(x)

}
}
