% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_generated_wrappers.R
\name{h2o4gpu.truncated_svd}
\alias{h2o4gpu.truncated_svd}
\title{Truncated Singular Value Decomposition (TruncatedSVD)}
\usage{
h2o4gpu.truncated_svd(n_components = 2L, algorithm = "power",
  n_iter = 100L, random_state = NULL, tol = 1e-05, verbose = FALSE,
  backend = "h2o4gpu", n_gpus = 1L, gpu_id = 0L)
}
\description{
Truncated Singular Value Decomposition (TruncatedSVD)
}
