% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{predict}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{use_condaenv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{reticulate}{\code{\link[reticulate]{use_condaenv}}}

  \item{stats}{\code{\link[stats]{predict}}}
}}

