% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{h2o4gpu.kmeans}
\alias{h2o4gpu.kmeans}
\title{K-Means clustering}
\usage{
h2o4gpu.kmeans(n_clusters = 8L, init = "k-means++", n_init = 1L,
  max_iter = 300L, tol = 1e-04, precompute_distances = "auto",
  verbose = 0, random_state = NULL, copy_x = TRUE, n_jobs = 1,
  algorithm = "auto", gpu_id = 0, n_gpus = -1, do_checks = 1)
}
\description{
K-Means clustering
}
