% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict.h2o4gpu_model}
\alias{predict.h2o4gpu_model}
\title{Make Predictions using Trained H2O4GPU Estimator}
\usage{
\method{predict}{h2o4gpu_model}(object, x, ...)
}
\arguments{
\item{x}{The new data where each column represents a different predictor variable to
be used in generating predictions.}

\item{...}{Additional arguments (unused for now).}
}
\description{
This function makes predictions from new data using a trained H2O4GPU model.
}
