% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_data}
\alias{prep_data}
\title{Prep data for h2o4gpu}
\usage{
prep_data(data_path = "", response = "", output_csv_name = "data.csv")
}
\arguments{
\item{data_path}{Path to data}

\item{response}{Response column as a string}

\item{output_csv_name}{Name of save csv file}
}
\description{
Prep data for h2o4gpu
}
