% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{transform.h2o4gpu_model}
\alias{transform.h2o4gpu_model}
\title{Transform a Dataset using Trained H2O4GPU Estimator}
\usage{
\method{transform}{h2o4gpu_model}(object, x, ...)
}
\arguments{
\item{object}{The h2o4gpu model object}

\item{x}{The new data where each column represents a different predictor variable to
be used in generating predictions.}

\item{...}{Additional arguments (unused for now).}
}
\description{
This function transforms the given new data using a trained H2O4GPU model.
}
